added gitignore, refactored the project
