+++ app.py
@@ -1,88 +1,27 @@
 from flask import Flask, request
 from flask_ask import Ask, statement, question, session
-from location_status import generate_status, generate_place_info
 from menu import get_menu
+from utility import get_custom_value, generate_place_info, get_open_places, statement_helper, read_menu
+import responses
 
 app = Flask(__name__)
 
 ask = Ask(app, "/ub_dining")
 
-DEFAULT_LOCATION = "North Campus Academic Buildings"
-ADDITIONAL_QUESTION = "Can I help you with something else?"
-ERROR_STATEMENT = "Sorry I don't know what you mean!"
-
-
-def get_open_places(location, date=None):
-    open_places = generate_status(location, date)
-    alexa_out = []
-    for place in open_places:
-        duration = open_places[place].replace('-', 'to')
-        s = f"{place} from {duration}. "
-        alexa_out.append(s)
-    return alexa_out
-
-
-def get_custom_value(content, intent, id_value=False):
-    try:
-        if id_value:
-            return content["request"]["intent"]["slots"][intent]["resolutions"][
-                "resolutionsPerAuthority"][0]["values"][0]["value"]["id"]
-        else:
-            return content["request"]["intent"]["slots"][intent]["resolutions"][
-                "resolutionsPerAuthority"][0]["values"][0]["value"]["name"]
-    except:
-        return None
-
-
-def statement_helper(plcaes, location):
-    places_msg = ""
-    if len(plcaes) == 1:
-        places_msg = f"The only food place open in {location} is {plcaes[0]}"
-    elif len(plcaes) == 0:
-        places_msg = f"There are no food places open in {location}! "
-    else:
-        msg = ""
-        for places_text in plcaes:
-            msg += places_text
-        places_msg = f"The food places open in {location} are {msg}"
-    places_msg += ADDITIONAL_QUESTION
-    return places_msg
-
-
-def read_menu(menu, place, time=None):
-    if menu == "Closed":
-        return f"{place} is Closed! "
-    else:
-        if time is None or len(menu) > 1:
-            menu_msg = f"The menu for {place} is..."
-            for meal_time in menu:
-                menu_msg += f"{meal_time} menu..."
-                for item in menu[meal_time]:
-                    menu_msg += f"{item}, "
-        else:
-            menu_msg = f"The {time} menu for {place} is "
-            for item in menu:
-                menu_msg += f"{item}, "
-        return menu_msg
-
 
 @app.route('/')
 def homepage():
-    return "UB Dining Alexa skill details will be here soon!"
+    return responses.HOMEPAGE_TEXT
 
 
 @ask.launch
 def start_skill():
-    welcome_message = 'Hello, I will help you with information about dining places in UB. To know more about what I can do just say help!'
-    return question(welcome_message)
+    return question(responses.WELCOME_MESSAGE)
 
 # Three things: Ask what is open at any day; Ask if that a particular location is open or not; Ask the menu for a dining center.
 @ask.intent("AMAZON.HelpIntent")
 def help():
-    help_text = "You can ask about what is open in UB..." + \
-        "You can also ask if a particular dining location is open or not and you can ask for the menu for a dining center in UB..." + \
-        "For sample commands look at the alexa skills page..." + ADDITIONAL_QUESTION
-    return question(help_text)
+    return question(responses.HELP_RESPONSE)
 
 
 @ask.intent("AMAZON.CancelIntent")
@@ -99,11 +38,11 @@ def open_by_date(time):
         content = request.json
         location = get_custom_value(content, "location")
         if location is None:
-            location = DEFAULT_LOCATION
+            location = responses.DEFAULT_LOCATION
         places = get_open_places(location, time)
         return question(statement_helper(places, location))
     except:
-        msg = ERROR_STATEMENT+ADDITIONAL_QUESTION
+        msg = responses.ERROR_STATEMENT+responses.ADDITIONAL_QUESTION
         return question(msg)
 
 
@@ -121,10 +60,10 @@ def check_place_open(time):
         else:
             duration = place_info.replace("-", "to")
             msg = f"{place} is open from {duration}..."
-        msg += ADDITIONAL_QUESTION
+        msg += responses.ADDITIONAL_QUESTION
         return question(msg)
     except:
-        msg = ERROR_STATEMENT+ADDITIONAL_QUESTION
+        msg = responses.ERROR_STATEMENT
         return question(msg)
 
 
@@ -139,17 +78,16 @@ def menu(time):
             raise Exception("Invalid dining location id")
         menu = get_menu(dining_place_id, time, meal_time)
         out_msg = read_menu(menu, dining_place, meal_time)
-        out_msg += ADDITIONAL_QUESTION
+        out_msg += responses.ADDITIONAL_QUESTION
         return question(out_msg)
     except Exception as ex:
-        msg = ERROR_STATEMENT+ADDITIONAL_QUESTION
+        msg = responses.ERROR_STATEMENT
         return question(msg)
 
 
 @ask.intent("AMAZON.FallbackIntent")
 def default_fallback():
-    msg = ERROR_STATEMENT + ADDITIONAL_QUESTION
-    return question(msg)
+    return question(responses.ERROR_STATEMENT)
 
 
 if __name__ == "__main__":
