+++ responses.py
@@ -1,0 +1,9 @@
+
+HOMEPAGE_TEXT = "UB Dining Alexa skill details will be here soon!"
+DEFAULT_LOCATION = "North Campus Academic Buildings"
+ADDITIONAL_QUESTION = "Can I help you with something else?"
+ERROR_STATEMENT = f"Sorry I don't know what you mean! {ADDITIONAL_QUESTION}"
+WELCOME_MESSAGE = "Hello, I will help you with information about dining places in UB. To know more about what I can do just say help!"
+HELP_RESPONSE = f"You can ask about what is open in UB..." + \
+    "You can also ask if a particular dining location is open or not and you can ask for the menu for a dining center in UB..." + \
+    "For sample commands look at the alexa skills page...{ADDITIONAL_QUESTION}"
